cmake windows qt5 msvc 2019

+++ src/core/p_calc_stack.cpp
@@ -71,7 +71,7 @@ void PCalcStack::prepareToExecute()
  * Zwraca rozmiar stosu.
  * @return rozmiar stosu
  */
-int PCalcStack::size()
+size_t PCalcStack::size()
 {
 	return values.size();
 }
